2025_tinkoff_ctf/web/caparet/sources/pyproject.toml
[project]
name = "caparet"
version = "0.1.0"
description = "A simple web app to book tickets for a theatre play"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "uvicorn>=0.29.0",
    "jinja2>=3.1.3",
    "pycryptodome>=3.20.0",
    "gunicorn>=22.0.0",
    "requests>=2.32.3",
]
